Botardo-Samsung/requirements.txt
#Python 3.10

certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==25.0
pillow==11.0.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
sympy==1.13.3
tokenizers==0.21.4
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.54.1
typing_extensions==4.14.1
urllib3==2.5.0
pyTelegramBotAPI
pytelegrambotapi
python-dotenv
thefuzz
PyAudio
SpeechRecognition

